requirements.txt
Flask==2.0.1
python-dotenv==0.19.0
gunicorn==20.1.0
Werkzeug==2.0.1
Flask-WTF==0.15.1
Flask-Limiter==1.4
python-docx==0.8.11
apscheduler==3.7.0